app/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.120.2
greenlet==3.2.4
h11==0.16.0
idna==3.11
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
psycopg2==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
